-Removed All eniteties and models and. -Created Bike Entity with the controller and views and models and finished it. -Delete Modal has a bug needs to be fixed.

+++ Aurora_Project/Controllers/.vshistory/BikesController.cs/2023-08-24_04_35_35_992.cs
@@ -4,6 +4,7 @@ using Aurora_Project.Data;
 using Aurora_Project.Data.Entities;
 using AutoMapper;
 using Aurora_Project.Models.Bikes;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace Aurora_Project.Controllers
 {
@@ -25,88 +26,78 @@ namespace Aurora_Project.Controllers
         public async Task<IActionResult> Index()
         {
             var bikes = await _context
-                                     .Bikes
-                                     .ToListAsync();
+                                      .Bikes
+                                      .ToListAsync();
 
-            var bikeVMs = _mapper.Map<List<Bike>, List<BikeIndexViewModel>>(bikes);
+            var bikesVM = _mapper.Map<List<Bike>, List<BikeViewModel>>(bikes);
 
-
-            return View(bikeVMs);
+            return View(bikesVM);
         }
 
 
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Bikes == null)
             {
                 return NotFound();
             }
 
-            var bike = await _context
-                                    .Bikes
-                                    .FirstOrDefaultAsync(m => m.Id == id);
+            var bike = await _context.Bikes
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (bike == null)
             {
                 return NotFound();
             }
 
-            var bikeVM = _mapper.Map<Bike, BikeDetailsViewModel>(bike);
-
-
-            return View(bikeVM);
+            return View(bike);
         }
 
 
         public IActionResult Create()
         {
-            return View();
+            var bikeVM = new CreateUpdateBikeViewModel();
+
+            bikeVM.BikeTypesSelectList = new SelectList(_context.BikeTypes, "Id","Type");
+
+            return View(bikeVM);
         }
 
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Brand,Model,Year,Color")] BikeCreateUpdateViewModel bikeVM)
+        public async Task<IActionResult> Create([Bind("Id,Brand,Model,Year,BikeType,Color")] Bike bike)
         {
             if (ModelState.IsValid)
             {
-                var bike = _mapper.Map<BikeCreateUpdateViewModel, Bike>(bikeVM);
-
                 _context.Add(bike);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(bikeVM);
+            return View(bike);
         }
 
 
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Bikes == null)
             {
                 return NotFound();
             }
 
-            var bike = await _context
-                                    .Bikes
-                                    .FindAsync(id);
-
-
+            var bike = await _context.Bikes.FindAsync(id);
             if (bike == null)
             {
                 return NotFound();
             }
-
-            var bikeVM = _mapper.Map<Bike, BikeCreateUpdateViewModel>(bike);
-
-            return View(bikeVM);
+            return View(bike);
         }
 
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Brand,Model,Year,Color")] BikeCreateUpdateViewModel bikeVM)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Brand,Model,Year,BikeType,Color")] Bike bike)
         {
-            if (id != bikeVM.Id)
+            if (id != bike.Id)
             {
                 return NotFound();
             }
@@ -115,14 +106,12 @@ namespace Aurora_Project.Controllers
             {
                 try
                 {
-                    var bike = _mapper.Map<BikeCreateUpdateViewModel, Bike>(bikeVM);
-
                     _context.Update(bike);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!BikeExists(bikeVM.Id))
+                    if (!BikeExists(bike.Id))
                     {
                         return NotFound();
                     }
@@ -133,17 +122,35 @@ namespace Aurora_Project.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            return View(bikeVM);
+            return View(bike);
         }
 
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null || _context.Bikes == null)
+            {
+                return NotFound();
+            }
+
+            var bike = await _context.Bikes
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (bike == null)
+            {
+                return NotFound();
+            }
+
+            return View(bike);
+        }
 
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Bikes == null)
+            {
+                return Problem("Entity set 'ApplicationDbContext.Bikes'  is null.");
+            }
             var bike = await _context.Bikes.FindAsync(id);
-
-
             if (bike != null)
             {
                 _context.Bikes.Remove(bike);
@@ -151,8 +158,6 @@ namespace Aurora_Project.Controllers
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-
-
         }
         #endregion
