+++ Aurora_Project/Controllers/.vshistory/BikesController.cs/2023-08-26_21_43_26_681.cs
@@ -1,112 +1,96 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Aurora_Project.Data;
 using Aurora_Project.Data.Entities;
-using AutoMapper;
-using Aurora_Project.Models.Bikes;
 
 namespace Aurora_Project.Controllers
 {
     public class BikesController : Controller
     {
-        #region Data & Constructors
         private readonly ApplicationDbContext _context;
-        private readonly IMapper _mapper;
 
-        public BikesController(ApplicationDbContext context, IMapper mapper)
+        public BikesController(ApplicationDbContext context)
         {
             _context = context;
-            _mapper = mapper;
         }
-        #endregion
 
-
-        #region Actions
+        // GET: Bikes
         public async Task<IActionResult> Index()
         {
-            var bikes = await _context
-                                     .Bikes
-                                     .ToListAsync();
-
-            var bikeVMs = _mapper.Map<List<Bike>, List<BikeIndexViewModel>>(bikes);
-
-
-            return View(bikeVMs);
+              return _context.Bikes != null ? 
+                          View(await _context.Bikes.ToListAsync()) :
+                          Problem("Entity set 'ApplicationDbContext.Bikes'  is null.");
         }
 
-
+        // GET: Bikes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Bikes == null)
             {
                 return NotFound();
             }
 
-            var bike = await _context
-                                    .Bikes
-                                    .FirstOrDefaultAsync(m => m.Id == id);
+            var bike = await _context.Bikes
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (bike == null)
             {
                 return NotFound();
             }
 
-            var bikeVM = _mapper.Map<Bike, BikeDetailsViewModel>(bike);
-
-
-            return View(bikeVM);
+            return View(bike);
         }
 
-
+        // GET: Bikes/Create
         public IActionResult Create()
         {
             return View();
         }
 
-
+        // POST: Bikes/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Brand,Model,Year,Color")] BikeCreateUpdateViewModel bikeVM)
+        public async Task<IActionResult> Create([Bind("Id,Brand,Model,Year,Color")] Bike bike)
         {
             if (ModelState.IsValid)
             {
-                var bike = _mapper.Map<BikeCreateUpdateViewModel, Bike>(bikeVM);
-
                 _context.Add(bike);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(bikeVM);
+            return View(bike);
         }
 
-
+        // GET: Bikes/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Bikes == null)
             {
                 return NotFound();
             }
 
-            var bike = await _context
-                                    .Bikes
-                                    .FindAsync(id);
-
-
+            var bike = await _context.Bikes.FindAsync(id);
             if (bike == null)
             {
                 return NotFound();
             }
-
-            var bikeVM = _mapper.Map<Bike, BikeCreateUpdateViewModel>(bike);
-
-            return View(bikeVM);
+            return View(bike);
         }
 
-
+        // POST: Bikes/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Brand,Model,Year,Color")] BikeCreateUpdateViewModel bikeVM)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Brand,Model,Year,Color")] Bike bike)
         {
-            if (id != bikeVM.Id)
+            if (id != bike.Id)
             {
                 return NotFound();
             }
@@ -115,14 +99,12 @@ namespace Aurora_Project.Controllers
             {
                 try
                 {
-                    var bike = _mapper.Map<BikeCreateUpdateViewModel, Bike>(bikeVM);
-
                     _context.Update(bike);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!BikeExists(bikeVM.Id))
+                    if (!BikeExists(bike.Id))
                     {
                         return NotFound();
                     }
@@ -133,35 +115,49 @@ namespace Aurora_Project.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            return View(bikeVM);
+            return View(bike);
         }
 
+        // GET: Bikes/Delete/5
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null || _context.Bikes == null)
+            {
+                return NotFound();
+            }
 
+            var bike = await _context.Bikes
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (bike == null)
+            {
+                return NotFound();
+            }
+
+            return View(bike);
+        }
+
+        // POST: Bikes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Bikes == null)
+            {
+                return Problem("Entity set 'ApplicationDbContext.Bikes'  is null.");
+            }
             var bike = await _context.Bikes.FindAsync(id);
-
-
             if (bike != null)
             {
                 _context.Bikes.Remove(bike);
             }
-
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-
-
         }
-        #endregion
 
-
-        #region Private Methods
         private bool BikeExists(int id)
         {
-            return (_context.Bikes?.Any(e => e.Id == id)).GetValueOrDefault();
-        } 
-        #endregion
+          return (_context.Bikes?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
     }
 }
